pyproject.toml
[tool.poetry]
name = "sql-chemistry"
version = "0.1.0"
description = ""
authors = ["dmkael <tanmaar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
